requirements.txt
fastapi[standard]
cloudscraper
urllib3
beautifulsoup4
python-dotenv